fix_InternalStorage
