+++ python/paddle/distributed/fleet/meta_optimizers/dygraph_optimizer/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 from .hybrid_parallel_optimizer import HybridParallelOptimizer
 from .hybrid_parallel_gradscaler import HybridParallelGradScaler
-# from .dygraph_sharding_optimizer import DygraphShardingOptimizer
+from .dygraph_sharding_optimizer import DygraphShardingOptimizer
 
 __all__ = []
