+++ python/paddle/distributed/fleet/meta_parallel/sharding/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .sharding_utils import GpuInfo
+from .sharding_utils import Type, device_guard
