+++ python/paddle/fluid/tests/unittests/dygraph_sharding_optimizer_stage2.py
@@ -24,9 +24,8 @@ import paddle.fluid as fluid
 from paddle.fluid.dygraph.nn import Linear
 from paddle.distributed import fleet
 
-from paddle.distributed.fleet.meta_parallel.sharding.sharding_utils import GpuInfo
 from paddle.distributed.fleet.utils.internal_storage import GradStorage
-from paddle.distributed.fleet.meta_optimizers.dygraph_optimizer.dygraph_sharding_optimizer import ShardingOptimizerStage2
+from paddle.distributed.fleet.meta_optimizers.dygraph_optimizer.sharding_optimizer_stage2 import ShardingOptimizerStage2
 
 base_lr = 0.1
 momentum_rate = 0.9
@@ -69,7 +68,6 @@ def optimizer_setting(parameter_list=None):
     return optimizer
 
 
-@GpuInfo
 def train_mlp():
     fleet.init(is_collective=True)
     group = paddle.distributed.new_group([0, 1])
